修复WlanQueryInterface接口方法映射

+++ src/main/java/net/java/dev/wlanapi/WlanApi.java
@@ -230,7 +230,7 @@ public interface WlanApi extends Library
     public int WlanQueryInterface(
     		Pointer hClientHandle,
     		Pointer pInterfaceGuid,
-    		long OpCode,
+    		int OpCode,
     		PointerByReference pReserved,
     		IntByReference pdwDataSize,
     		PointerByReference ppData,
